deeplearning.ai/langchain-for-llm-app-dev-course/01-structured-output-examples/requirements.txt
langchain==0.3.18
langchain-core==0.3.28
langchain-openai==0.2.14
python-dotenv==1.0.0
pydantic==2.10.6
